orders/pyproject.toml
[project]
name = "Review"
version = "1.0.0"
description = "Service for coffy shop reviws"
authors = [
    {name = "Jan Kolomazník", email = "jan.kolomaznik@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.27.0",
    "pika>=1.3.2",
    "pydantic>=2.4.2",
    "sqlalchemy>=2.0.22",
    "psycopg2>=2.9.9",
    "grpcio-tools>=1.60.0",
    "grpcio>=1.59.3",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload"

